Develop (#54)
* feat: add model prisma

* chore: add schema prisma

* feat: add shadcn framwork

* build(web): installed dropwdown, nav-menu, separator, react-icons, & embla-carousel

* build(web): added parse thousand func on utils.ts. Added 1 path for route groups on tailwind config

* style(web): added Footer & Header on RootLayout at (home) route

* style(web): created HeroCarousel.tsx from shadcn carousel components

* style(web): created BestSeller.tsx, Banner.tsx, & Campaign.tsx

* style(web): added HeaderDropdown.tsx for screen smaller than desktop

* style(web): putting HeroCarousel, BestSeller, Banner, and Campaign on page.tsx on (home) route

* feat: add to cart

* fix: debugging

* fix: add to cart

* fix: debug

* Update .env.development

* feat: category management

* fix: test

* fix: test

* feat: update cart

* style(web): hide HeaderDropdown on md & moved all landing-page-related to _components at (home)

* fix: fix

* fix: disabled user

* build: installed firebase & its sso login function. Added useStorage hook

* build(api): modified schema & re-migrated prisma schema. Added database_url, key_jwt, mailer on .env

* build(api): install bcrypt, handlebars, jsonwebtoken, uuid & nodemailer. Add template => nodemailer

* feat(web): added feature user register & user social login

* feat(api): added createUser controller and router

* feat(api): added setupUser & verifyUser controller

* feat(web): created slug page for setup user account & verify on /verify route

* fix: deleted import from account.controller

* fix: added .env.development on .gitignore

* fix: prisma migration, removed uuid duplicate & conditional on deleteCartItems

* fix: changed IDRformatter function to formatToIDR. Moved admins route to (dashboard)

* feat: added login user feature

* fix: user data from social login will be saved to mysql

* build(web): installed jwt-decode package

* feat(web): add function & useEffect if token is expired

* feat(api): setup router & controller to handle refresh token

* style(web): styled user verification page

* fix(web): changed img tag to Image from next on components on homepage & auth & verification page

* refactor(web): transformed background on verify page BrickPhotoBackground.tsx component

* feat: wear-25

* feat: wear 25

* feat: forgot password & reset password

* fix: added User type for redux, added catalog route, initial middleware.ts

* fix: added ProfileDropdown for logged user, ForgotPassForm, /manage-address route,  getUserById cont

* feat: create order

* build: added profileSlice & removeImageFromStorage on useStorage hooks

* style: made UserMenu, EditProfileForm, & its page.tsx responsive

* fix: schema warehouse update

* feat: wear-25

* fix: fix merge

* fix: fix merge week 2

* feat: stock not finished

* fix: fixed bug on login using google

* fix: authorization on middleware.ts

* feat: createAdmin, verify admin, & login to dashboard feature

* feat: user can change photo profile

* fix: middleware, & navbar position & added logout function

* fix(style): condition on LoadingButton & moved InputPhotoProfile component

* feat: edit user profile, re-verify when email is changed, reject email change for SSO user

* feat: manage address still on progress

* fix: clear WEAR-18 WEAR-31

* fix: fix bug

* feat: added getAdminClientSide & commented duplicate searchbar

* feat: stash

* style: fixed Header with its userLogged state & fixed Wrapper heading font size

* fix: user photo profile preview & added delete photo profile feature

* feat: wear-26 & wear-9

* feat: wear-26 & wear-9

* feat: add new feature for WEAR-28

* fix: test only

* feat: wear-10

* fix: clear WEAR-28

* fix: delete invoice

* fix: add public/pdf

* fix: set up WEAR-30

* feat: add WEAR-30 & create mutation stock

* fix: debug for npm run build

* fix: fix order detail page

* feat: wear-27 & wear-33

* fix: wear-27 & wear-33

* feat: manage address (user), manage warehouse, user, admin (super admin)

* feat: manage address (user), manage warehouse, user, admin (super admin)

* fix: fixed schema & apos error

* fix: unsafe type on admin & user. Fix user, admin, & warehouse fetching on admin dashboard for table

* fix: fe feature 2

* fix: update create mutation

* fix: added db seeding for users & admins

* fix: added no-cache to getRequest function, & seeding for warehouses

* fix: fix chekout & order page

* fix: authorization,  protect admin route from user

* fix: added isActive column on Warehouse table & its seeding data

* fix: actions id column, added checkbox, & added export to excel function on the table

* style: fixed dashboard admin sidebar to be h-screen & added overflow-scroll on DashboardWrapper

* fix: update database url

* fix: add filter order by date and fix pagination order bug

* fix: add favicon

* fix: fix add to cart

* feat: added assign to & dismiss admin from warehouse feature

* fix: changed literal api-url to process.env

* feat: deactivate & reactivate warehouse feature

* build: added product seeding

* fix: warehouse seeding

* fix: manage warehouse & manage user address

* fix: edit deploy config

* fix: edit config deploy

* feat: wear-32

* feat: wear-32

* feat: wear-32

* fix: productCategory seeding

* fix: get categories for Header, fixed Search, dropdown on header, & dynamic BestSeller comp

* style: added DashboardHeaderPhoto, ExcelButton, added header on users, admins, & warehouse

* fix: removed getAllWarehouses controller

* fix: removed facebook login

* fix: fix filter order by date

* fix: edit github workflow

* fix: edit config json

* fix: develop

* fix: products

* fix: env production

* fix: package.json

* fix: prod controller

* fix: test

* fix: fix transaction mutation

* fix: get catalogs

---------

+++ apps/api/src/services/product/edit.action.ts
@@ -13,7 +13,7 @@ export async function editProduct(req: Request, res: Response) {
     }
     await prisma.$transaction(async (tx) => {
         const { slug } = req.params;
-        const { name, description, price, thumbnailURL, additionalURL, additionalDelete, colorVariantEdit, colorVariantNew, colorVariantDelete, categoryData } = req.body;
+        const { name, isVisible, description, price, thumbnailURL, additionalURL, additionalDelete, colorVariantEdit, colorVariantNew, colorVariantDelete, categoryData } = req.body;
         const symbolRegex = /[^a-zA-Z0-9\s-]/;
         if (name.length > 30 || name.length < 2 || name.trim().length === 0 || symbolRegex.test(name)) throw "Invalid name input."
         if (description.length > 500 || description.length < 15 || description.trim().length === 0) throw "Invalid description input."
@@ -34,6 +34,7 @@ export async function editProduct(req: Request, res: Response) {
                 name,
                 description,
                 price,
+                isActive: isVisible,
                 slug: name.toLowerCase().replaceAll(" ", "-"),
                 categoryID: productCategory!.id,
                 updatedAt: new Date()
@@ -139,41 +140,43 @@ export async function editProduct(req: Request, res: Response) {
                             productVariantID: colorVariantDelete[i]
                         }
                     })   
-                    const stockLog = await tx.stockMutation.create({
-                        data: {
-                            id: uuidv4(),
-                            warehouseID: wareHouseList[k].id,
-                            type: 'DELETE',
-                        }
-                    })
-                    for (let c = 0; c<currentStock.length; c++) {
-                        await tx.stockMutationItem.create({
+                    if (currentStock.length > 0) {
+                        const stockLog = await tx.stockMutation.create({
                             data: {
                                 id: uuidv4(),
-                                quantity: currentStock[c].stock,
-                                stockMutationID: stockLog.id,
-                                warehouseProductID: currentStock[c].id
+                                warehouseID: wareHouseList[k].id,
+                                type: 'DELETE',
                             }
                         })
-                        await tx.warehouseProduct.updateMany({
-                            where: {
-                                productVariantID: colorVariantDelete[i],
-                                warehouseID: wareHouseList[k].id
-                            },
-                            data: {
-                                isDelete: true,
-                                stock: 0
-                            }
-                        })
-                        await tx.productVariant.update({
-                            where: {
-                                id: colorVariantDelete[i]
-                            },
-                            data: {
-                                isDeleted: true
-                            }
-                        })
+                        for (let c = 0; c<currentStock.length; c++) {
+                            await tx.stockMutationItem.create({
+                                data: {
+                                    id: uuidv4(),
+                                    quantity: currentStock[c].stock,
+                                    stockMutationID: stockLog.id,
+                                    warehouseProductID: currentStock[c].id
+                                }
+                            })
+                            await tx.warehouseProduct.updateMany({
+                                where: {
+                                    productVariantID: colorVariantDelete[i],
+                                    warehouseID: wareHouseList[k].id
+                                },
+                                data: {
+                                    isDelete: true,
+                                    stock: 0
+                                }
+                            })
+                        }
                     }
+                    await tx.productVariant.update({
+                        where: {
+                            id: colorVariantDelete[i]
+                        },
+                        data: {
+                            isDeleted: true
+                        }
+                    })
                 }
